[Refactor] Rename restaurant detail component
